+++ bradata/agencias/__init__.py
@@ -1,0 +1,11 @@
+import bradata.agencias.infraero
+"""
+Gets content from infraero website. It provides a mapping to content types.
+You can initialize this submodule by:
+    import bradata.agencias.infraero as infraero
+    infraero <do something>
+or
+    import bradata
+    bradata.agencias.infraero.<do something>
+and you'll be ready to use the API on your Python project.
+"""